变更 id

+++ FMO.Automation/GatherDailyFromMail/GzMailAttachInfo.cs
@@ -8,7 +8,7 @@ public class GzMailAttachInfo
 
     public int? FundId { get; set; }
 
-    public int? DailyId { get; set; }
+    public long? DailyId { get; set; }
 
     [BsonIgnore]
     public bool IsMapped => FundId is not null && DailyId is not null;
